pyproject.toml
[project]
name = "context9"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Pawel Konior", email = "pawel.konior@gmail.com" }
]
requires-python = ">=3.14.4"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "fastapi>=0.136.1",
    "fastmcp>=3.2.4",
    "httpx>=0.28.1",
    "lxml>=6.1.0",
    "pydantic-settings>=2.14.0",
    "qdrant-client>=1.17.1",
    "rich>=15.0.0",
    "typer>=0.24.2",
    "uvicorn[standard]>=0.46.0",
]

[project.scripts]
context9 = "context9:main"
context9-embedder = "context9.embedder_api:run"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pytest>=9.0.3",
    "pytest-asyncio>=1.3.0",
    "pytest-cov==7.0.0",
    "pytest-mock==3.15.1",
    "pytest-randomly==4.0.1",
    "pytest-xdist==3.8.0",
    "pytest-timeout==2.4.0",
    "pre-commit==4.5.1",
    "respx>=0.23.1",
    "ruff>=0.15.12",
    "sphinx>=9.1.0",
    "ty>=0.0.32",
]

[tool.coverage.run]
source = ["src/context9"]
branch = true
parallel = true
relative_files = true

[tool.coverage.report]
show_missing = true
skip_covered = true
fail_under = 100
